Enable strict typescript
Thank you very much for this awesome seed, we are using quite heavily.

There are two things we always change in our projects:

Enabling strict typescript. We think this is the best pratice and doing
that from early on makes everything simpler in the long run.

Removing the "do not install node-modules if the directory already
exists" (see thee change in ui-build.sbt). With our CI setup this would
always mean that Jenkins would not update node modules if there was a
previous run which used the same workspace and already installed
modules.